Basic User Premission for Posts

+++ src/posts/views.py
@@ -1,7 +1,7 @@
 from urllib.parse import quote_plus
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 
 # Create your views here.
@@ -10,6 +10,8 @@ from .models import Post
 
 
 def post_create(request):
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	form = PostForm(request.POST or None, request.FILES or None)
 	if form.is_valid():
 		instance = form.save(commit=False)
@@ -67,6 +69,8 @@ def post_list(request): #list
 	#return HttpResponse("<h1>list</h1>")
 
 def post_update(request, slug=None):
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	form = PostForm(request.POST or None, request.FILES or None, instance=instance)
 	if form.is_valid():
@@ -85,7 +89,9 @@ def post_update(request, slug=None):
 	return render(request, "post_form.html", context)
 
 
-def post_delete(request, slug=None):
+def post_delete(request, slug=None):\
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	instance.delete()
 	messages.success(request, "Successfully Deleted")
